+++ packages/utils/src/config/index.ts
@@ -1,0 +1,1 @@
+export { default as normalizeHints } from './normalize-hints';
